refactor: Regroup auth routes under users
