+++ server/src/router/routes/userRoutes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { register, login } from '../../controllers/authController';
+import { register, login } from '../../controllers/userController';
 import { validateRegistration } from '../../middleware/validation';
 
 const router = Router();
